+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+# v3.8.0
+- [BP-1462](https://movai.atlassian.net/browse/BP-1462): Translation implementation
+  - Singleton - lock only on first instantiation
+  - Add `subscribe_by_args_decoded` to simplify subscription
+
+# v3.7.1
+- [BP-1470](https://movai.atlassian.net/browse/BP-1470): Move DEFAULT_LANGUAGE to central location
+
+# v3.7.0
+- [BP-1456](https://movai.atlassian.net/browse/BP-1456): Add support for user language configuration
+  - Introduced `language` attribute for `BaseUser` and `InternalUser` models.
+  - Added validation for `language` using `VALID_LANGUAGES` from `movai_core_shared.consts`.
+  - Updated JSON schemas (`InternalUser.json` and `User.json`) to include the `Language` field.
+  - Added getter and setter for the `language` property in `BaseUser`.
+
+# v3.6.0
+- [BP-1472](https://movai.atlassian.net/browse/BP-1472): Import / remove / remove Translation scope
+
+# v3.5.1
+- [BP-1476](https://movai.atlassian.net/browse/BP-1476): tools.backup root-path (-r) parameter is not working
+
+# v3.5.0
+- [BP-1463](https://movai.atlassian.net/browse/BP-1463): Create translation scope
+
 # v3.4.1
 - Fix mobdata usage (missing 1 required positional argument: 'args')